// src/app/pimDetails/sections/documents/documentDetails/documentContract/documentSecurity/documentRights/DocumentRights.types.ts
export enum DocumentType {
  Leaflet = 'Leaflet',
  TitleDeed = 'TitleDeed',
  MunicipalCharges = 'MunicipalCharges',
  SignedDeedOfPurchase = 'SignedDeedOfPurchase',
  InformationForm = 'InformationForm',
  CadastralMap = 'CadastralMap',
  CadastralRegister = 'CadastralRegister',
  ServiceContract = 'ServiceContract',
  DeedOfDivision = 'DeedOfDivision',
}